[refactor]: 로딩 프로그래스 바를 설정해서 로그인/로그아웃/로딩 상태 표시
- Nav바에 AuthInitializer를 넣어 상태표시 UI적용
