+++ src/app/loading.tsx
@@ -1,0 +1,7 @@
+import Image from 'next/image';
+
+const Loading = () => {
+  return <Image src="/assets/loading.svg" alt="loading" width="40" height="40"></Image>;
+};
+
+export default Loading;
